选择过滤器扩展修改
